move dropout view tests
